requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
decorator==5.2.1
defusedxml==0.7.1
Django==5.2
django-allauth==65.7.0
django-crispy-forms==2.4
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
httpie==3.2.4
idna==3.10
json5==0.12.0
Markdown==3.8
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.4.3
mysqlclient==2.2.7
pillow==11.2.1
psycopg2-binary==2.9.10
Pygments==2.19.1
PyJWT==2.9.0
PySocks==1.7.1
python-dotenv==1.1.0
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
setuptools==78.1.0
sqlparse==0.5.3
urllib3==2.4.0
